# docs/api/generated/openapi.yaml
openapi: 3.0.0
info:
  title: Property Management API
  version: 0.0.0
...
paths:
...
  /api/v1/accounting/invoices:
    get:
      operationId: InvoicesApi_list
      description: List invoices for current tenant
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/Shared.PaginationQuery.page'
        - $ref: '#/components/parameters/Shared.PaginationQuery.limit'
        - $ref: '#/components/parameters/Shared.PaginationQuery.sortBy'
        - $ref: '#/components/parameters/Shared.PaginationQuery.sortOrder'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Accounting.Invoice'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
      tags:
        - Accounting
    post:
      operationId: InvoicesApi_create
      description: Create new invoice
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounting.Invoice'
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
      tags:
        - Accounting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Accounting.CreateInvoiceRequest'
  /api/v1/accounting/invoices/{id}:
    get:
      operationId: InvoicesApi_get
      description: Get invoice by ID
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Shared.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounting.Invoice'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '404':
          description: Not Found - Resource does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
      tags:
        - Accounting
    patch:
      operationId: InvoicesApi_update
      description: Update invoice
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Shared.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Accounting.Invoice'
        '400':
          description: Bad Request - Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '404':
          description: Not Found - Resource does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
      tags:
        - Accounting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Accounting.UpdateInvoiceRequest'
    delete:
      operationId: InvoicesApi_delete
      description: Delete invoice
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Shared.uuid'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '404':
          description: Not Found - Resource does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
      tags:
        - Accounting
  /api/v1/accounting/invoices/{id}/items:
    get:
      operationId: InvoicesApi_listItems
      description: List invoice items
      parameters:
        - name: Authorization
          in: header
          required: false
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Shared.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Accounting.InvoiceItem'
        '401':
          description: Unauthorized - Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
        '404':
          description: Not Found - Resource does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shared.ErrorResponse'
      tags:
        - Accounting
...
components:
  parameters:
    Shared.PaginationQuery.limit:
      name: limit
      in: query
      required: false
      description: Items per page
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
        default: 20
      explode: false
    Shared.PaginationQuery.page:
      name: page
      in: query
      required: false
      description: Page number (1-indexed)
      schema:
        type: integer
        format: int32
        minimum: 1
        default: 1
      explode: false
    Shared.PaginationQuery.sortBy:
      name: sortBy
      in: query
      required: false
      description: Sort field
      schema:
        type: string
      explode: false
    Shared.PaginationQuery.sortOrder:
      name: sortOrder
      in: query
      required: false
      description: Sort direction
      schema:
        type: string
        enum:
          - asc
          - desc
        default: asc
      explode: false
...
  schemas:
...
    Accounting.CreateInvoiceItem:
      type: object
      required:
        - description
        - qty
        - unitPrice
        - vatRate
      properties:
        description:
          type: string
          description: Description
        qty:
          type: number
          format: double
          description: Quantity
        unitPrice:
          type: number
          format: double
          description: Unit price excluding VAT
        vatRate:
          type: number
          format: double
          description: VAT rate percentage
        vatRateType:
          allOf:
            - $ref: '#/components/schemas/Accounting.VatRate'
          description: Optional named VAT rate (overrides vatRate if provided)
      description: Request to create a new invoice item
    Accounting.CreateInvoiceRequest:
      type: object
      required:
        - contactId
        - number
        - issueDate
        - dueDate
        - currency
        - items
      properties:
        contactId:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Contact identifier
        number:
          type: string
          description: Invoice number
        issueDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Issue date
        dueDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Due date
        taxableSupplyDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Taxable supply date
        currency:
          type: string
          description: Currency code (e.g., CZK)
        variableSymbol:
          type: string
          description: Variable symbol
        status:
          allOf:
            - $ref: '#/components/schemas/Accounting.InvoiceStatus'
          description: Initial status
        items:
          type: array
          items:
            $ref: '#/components/schemas/Accounting.CreateInvoiceItem'
          description: Invoice line items
        country:
          type: string
          description: Optional country code (e.g., 'CZ', 'SK') to resolve named VAT rates
      description: Request to create a new invoice
    Accounting.Invoice:
      type: object
      required:
        - id
        - tenantId
        - contactId
        - number
        - issueDate
        - dueDate
        - currency
        - totalAmount
        - baseAmount
        - vatAmount
        - status
        - paidAmount
        - createdAt
        - updatedAt
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Unique identifier
        tenantId:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Tenant identifier
        contactId:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Contact identifier
        number:
          type: string
          description: Invoice number
        issueDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Issue date
        dueDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Due date
        taxableSupplyDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Taxable supply date
        currency:
          type: string
          description: Currency code (e.g., CZK)
        totalAmount:
          type: number
          format: double
          description: Total amount including VAT
        baseAmount:
          type: number
          format: double
          description: Base amount excluding VAT
        vatAmount:
          type: number
          format: double
          description: VAT amount
        variableSymbol:
          type: string
          description: Variable symbol for payment identification
        status:
          allOf:
            - $ref: '#/components/schemas/Accounting.InvoiceStatus'
          description: Current status
        paidAmount:
          type: number
          format: double
          description: Amount already paid
        createdAt:
          allOf:
            - $ref: '#/components/schemas/Shared.dateTime'
          description: Creation timestamp
        updatedAt:
          allOf:
            - $ref: '#/components/schemas/Shared.dateTime'
          description: Last update timestamp
        createdBy:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: ID of user who created this entity
        updatedBy:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: ID of user who last updated this entity
      description: Native issued invoice
    Accounting.InvoiceItem:
      type: object
      required:
        - id
        - invoiceId
        - tenantId
        - description
        - qty
        - unitPrice
        - vatRate
        - totalAmount
        - baseAmount
        - vatAmount
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Unique identifier
        invoiceId:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Invoice identifier
        tenantId:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Tenant identifier
        description:
          type: string
          description: Description of the item
        qty:
          type: number
          format: double
          description: Quantity
        unitPrice:
          type: number
          format: double
          description: Unit price excluding VAT
        vatRate:
          type: number
          format: double
          description: VAT rate percentage
        totalAmount:
          type: number
          format: double
          description: Total amount including VAT
        baseAmount:
          type: number
          format: double
          description: Base amount excluding VAT
        vatAmount:
          type: number
          format: double
          description: VAT amount
      description: Line item for an invoice
    Accounting.InvoiceStatus:
      type: string
      enum:
        - draft
        - issued
        - paid
        - partially_paid
        - overdue
      description: Status of an issued invoice
...
    Accounting.UpdateInvoiceRequest:
      type: object
      properties:
        contactId:
          allOf:
            - $ref: '#/components/schemas/Shared.uuid'
          description: Contact identifier
        number:
          type: string
          description: Invoice number
        issueDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Issue date
        dueDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Due date
        taxableSupplyDate:
          allOf:
            - $ref: '#/components/schemas/Shared.date'
          description: Taxable supply date
        currency:
          type: string
          description: Currency code
        variableSymbol:
          type: string
          description: Variable symbol
        status:
          allOf:
            - $ref: '#/components/schemas/Accounting.InvoiceStatus'
          description: Status
        paidAmount:
          type: number
          format: double
          description: Paid amount
      description: Request to update an existing invoice
    Accounting.VatRate:
      type: string
      enum:
        - zero
        - reduced
        - first_reduced
        - standard
      description: VAT rate types for CZ/SK
...
    Shared.ErrorResponse:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: Error code for programmatic handling
        message:
          type: string
          description: Human-readable error message
        requestId:
          type: string
          description: Request ID for support/debugging
        details:
          type: array
          items:
            $ref: '#/components/schemas/Shared.ValidationError'
          description: Detailed validation errors
        timestamp:
          allOf:
            - $ref: '#/components/schemas/Shared.dateTime'
          description: ISO 8601 timestamp
      description: Standard error response
...
    Shared.ValidationError:
      type: object
      required:
        - field
        - message
        - code
      properties:
        field:
          type: string
          description: Field path (e.g., 'address.city')
        message:
          type: string
          description: Error message for this field
        code:
          type: string
          description: Error code
      description: Validation error detail
    Shared.date:
      type: string
      description: ISO 8601 date string (YYYY-MM-DD)
    Shared.dateTime:
      type: string
      description: ISO 8601 date-time string
...
    Shared.uuid:
      type: string
      description: Unique identifier (UUID v4)
...
servers:
  - url: http://localhost:8080
    description: Development
    variables: {}
  - url: https://api.property-management.example.com
    description: Production
    variables: {}